refactor: move stream-loss detection into kvutil
